platform/api/requirements.txt
# ========================================================
# Python Backend Dependencies (FastAPI Clerk + Workers)
# ========================================================

# Core API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0

# Data Ingestion (Project 1)
pandas>=2.2.0
openpyxl>=3.1.0  # For XLSX support

# Database
supabase>=2.4.0

# Messaging (RabbitMQ)
pika>=1.3.0

# Validation (Python 3.13 compatible)
pydantic>=2.10.0
email-validator>=2.1.0

# SMTP Worker (Phase 3)
aiosmtplib>=3.0.0
bleach>=6.1.0
